split RocketConfigs into RoCCAccelConfigs and MMIOAccelConfigs
